ecommerce-assistant/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.71.1
python-multipart==0.0.9
